+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.34.1] - 2026-07-08
+
 ### Fixed
 
 - **"daemon did not come up" when starting with a large saved workspace** —